INVT-9468 fails fast if action is not executed on git branch
To introduce versioning via git tags, we require the pipeline to be executed
on some git branch. If not, we have no way to read old tags and cannot introduce new
release tags.
In order to disallow the usage of this action without any git branches, we will now
fail before anything else is executed.