INVT-9643 change action's interface
We make branch an optional parameter, as it feels weird for clients to pass a branch name, when using this action to rollback a release.
For the same reason, I also decided to make source-directory an optional parameter.
I added error handling within the action, if the user does not perform a rollback and did not specify the parameters branch and source-directory.