INVT-10563 version all uncached files
The K8 CDN-proxy defines a list of files that are uncached.
We need to make sure that all of these files are versioned and rolled back to get consistent states after a rollback.
The list of files is defined here: https://github.com/leanix/cdn-proxy/blob/master/pkg/cdnproxy/bootstrapHttpHandler.go#L86