INVT-10720 refactor region and storage-container construction
The construction of the storage-container based on a region and the environment can be extracted into a separate function.
This makes this code reusable and reduces complexity of the main code.

Regions were also stored in a non-optimal way as the region objects have been constructed as { region: '...', short: '...' }.
However the data modeled for a region is { name: '...', short: '...' } where the context itself is given by the name of the object (e.g. region).