INVT-10721 add optional versioning step
If the user wants to version the current deployment, they can now use three additional input parameters that make this possible.
The deployment strategy calculates a release versioned based on existing release tags in the corresponding deployed branch.
This is the same strategy as previously used to create a versioned Docker container.

The versioning step is completly optional and clients, which do not provide any of the new input parameters will not notice any changes to their pipelines.