INVT-10721 add rollback functionality
Clients of the action can enter an optional rollback mode to go back to a previous version.
The responsibility of guaranteeing that the application is in a good state after the rollback is with client of the action.
Clients which do not employ versioning and rollback will not notice that the action provides this addtional functionality.