INVT-10721 extract script to fetch sas token
In the test workflow we provide integration tests that check if the deployments and rollback on the Azure File Share work as intended.
To access the file share, we need to fetch the SAS token.
To reduce the complexity of our test-workflow, we want to extract long and complicated shell scripts from it to provide some form of abstraction.