THRST-315 Upload uncached root files after all other files
Customers have complained about the frontend not being able to load a few times during the day.
This happens because we are running new releases at this time, and the storage-deploy-action has uploaded the index.html / main.js file before the assets that are referenced within these files.
The result is the customer experiencing a white page and the application only loading after like a minute of refreshing the page.

The fix is to upload all assets before uploading the uncached root files.