THRST-344 Minimise risk of chunk errors
Chunk errors still occur during deployment, because unchanged or non-fingerprinted assets are at points in time only partially uploaded.
If users access these assets during this time frame, they will receive a chunk error.
Our idea of prevention is to upload the release in a dedicated folder, and then afterwards we just move the files into the root directory which should be much faster and minimize the amount of chunk load errors that we receive.